refacor : @Valid 어노에티션 사용 및 리펙토링
