+++ src/main/java/com/missionx/questloggers/domain/auth/controller/AuthController.java
@@ -1,9 +1,10 @@
 package com.missionx.questloggers.domain.auth.controller;
 
+import com.missionx.questloggers.domain.auth.dto.SignupResponseDto;
 import com.missionx.questloggers.domain.auth.service.AuthService;
-import com.missionx.questloggers.domain.user.dto.SignupRequestDto;
-import com.missionx.questloggers.domain.user.dto.SignupResponseDto.SignupUserData;
+import com.missionx.questloggers.domain.auth.dto.SignupRequestDto;
 import com.missionx.questloggers.global.dto.ApiResponse;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,16 +14,15 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/auth")
+@RequestMapping("/api")
 @RequiredArgsConstructor
 public class AuthController {
 
     private final AuthService authService;
 
-    @PostMapping("/signup")
-    public ResponseEntity<ApiResponse<SignupUserData>> signup(@RequestBody SignupRequestDto signupRequestDto) {
-        SignupUserData response = authService.signup(signupRequestDto);
-        return ApiResponse.success(HttpStatus.CREATED, "회원가입이 완료되었습니다.", response);
+    @PostMapping("/auth/signup")
+    public ResponseEntity<ApiResponse<SignupResponseDto>> signup(@Valid @RequestBody SignupRequestDto signupRequestDto) {
+        SignupResponseDto signupResponseDto = authService.signup(signupRequestDto);
+        return ApiResponse.success(HttpStatus.CREATED, "회원가입이 완료되었습니다.", signupResponseDto);
     }
-
 }
