+++ src/main/java/com/missionx/questloggers/domain/auth/dto/SignupRequestDto.java
@@ -1,0 +1,21 @@
+package com.missionx.questloggers.domain.auth.dto;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Pattern;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class SignupRequestDto {
+    @Email(message = "이메일 형식이 올바르지 않습니다.")
+    @NotBlank(message = "이메일은 필수입니다.")
+    private String email;
+
+    @NotBlank(message = "비밀번호는 필수입니다.")
+    @Pattern(regexp = "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[~!@#$%^&*]).{8,}$", message = "비밀번호는 최소 8자, 대소문자, 숫자, 특수문자를 포함해야 합니다.")
+    private String password;
+
+    @NotBlank(message = "API 키는 필수입니다.")
+    private String apiKey;
+}
